requirements.txt
celery==3.1.23
coverage==4.2
Django==1.10.3
django-celery==3.1.17
djangorestframework==3.3.3
mysqlclient==1.3.9
Pillow==3.4.2
PyMySQL==0.7.9
